translator/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.4
colorama==0.4.4
coloredlogs==15.0.1
filelock==3.7.0
flatbuffers==1.12
grpcio==1.46.3
grpcio-tools==1.46.3
huggingface-hub==0.6.0
humanfriendly==10.0
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
nltk==3.7.0
numpy==1.22.4
onnx==1.11.0
onnxconverter-common==1.9.0
onnxruntime==1.11.1
onnxruntime-tools==1.7.0
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
psutil==5.9.1
py==1.11.0
py-cpuinfo==8.0.0
py3nvml==0.2.7
pyparsing==3.0.9
pyreadline3==3.4.1
pytest==7.1.2
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.16.0
termcolor==1.1.0
tf2onnx==1.10.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
urllib3==1.26.9
xmltodict==0.13.0
